requirements.txt
requests
websockets
psycopg2-binary==2.9.9